backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
pytest==8.3.4
pytest-mock==3.12.0
ruff==0.8.6
black==24.10.0
httpx==0.28.1
python-dotenv==1.0.1
openai==1.54.3
contentful-management==2.13.1
pre-commit==3.5.0
